replace job_id_t with uint64_t

+++ job.h
@@ -23,13 +23,14 @@
 
 #include <boost/shared_ptr.hpp>
 #include <string>
+#include <cstdint>
 
 namespace Beanstalkpp {
   
 class Client;
 class Job;
 
-typedef unsigned int job_id_t;
+typedef uint64_t job_id_t;
 typedef boost::shared_ptr<Job> job_p_t;
 
 /**
